refactor(#89): 도메인 보조 파일 위치 정돈
