+++ src/logs/logs.controller.ts
@@ -12,7 +12,7 @@ import {
 import { applyNoCacheHeaders } from '../common/http/response-helpers';
 import type { DungeonLogsPayload } from './dto/logs-response.dto';
 import { LogsService } from './logs.service';
-import { validateLogsQuery } from './logs-query.validator';
+import { validateLogsQuery } from './validators/logs-query.validator';
 import type { LogsQueryDto } from './dto/logs-query.dto';
 import { LogTypeEnum } from './logs.types';
