+++ src/ranking/validators/ranking-query.validator.ts
@@ -2,10 +2,13 @@ import {
   RANKING_DEFAULT_LIMIT,
   RANKING_MAX_LIMIT,
   RANKING_MIN_LIMIT,
-} from './ranking.constants';
-import { decodeRankingCursor, type RankingCursor } from './ranking-cursor.util';
-import { buildInvalidQueryException } from './ranking.errors';
-import type { RankingQueryDto } from './dto/ranking-query.dto';
+} from '../ranking.constants';
+import {
+  decodeRankingCursor,
+  type RankingCursor,
+} from '../utils/ranking-cursor.util';
+import { buildInvalidQueryException } from '../ranking.errors';
+import type { RankingQueryDto } from '../dto/ranking-query.dto';
 
 export type ValidatedRankingQuery = {
   limit: number;
